fix date format trouble use type string and add error to the controller removeById

+++ controllers/movies/removeById.js
@@ -3,6 +3,7 @@ const { Movie } = require('../../models/movie');
 const removeById = async (req, res) => {
   const { movieId } = req.params;
   const removedMovie = await Movie.findById(movieId);
+  if (!removedMovie) throw RequestError(404, 'Not found');
   await Movie.findByIdAndRemove(movieId);
   res.status(200).json({ message: 'Movie was removed', removedMovie }); // status 200 - because 204 do not return a message
 };
